+++ packages/outbox/src/test/pushers/jobs/jobs-outbox.pusher.int.spec.ts
@@ -46,7 +46,7 @@ describe('JobsOutboxPusher (Integration)', () => {
 
       expect(job).toBeDefined();
       expect(job?.name).toBe('test.job');
-      expect(job?.data).toEqual({ foo: 'bar' });
+      expect(job?.data).toEqual({ payload: { foo: 'bar' }, emitter: 'database-tests' });
     } finally {
       await inspectorQueue.close();
     }